refactor: unify into single drua binary with server/client/cli crates
Split into three crates with clear responsibilities:
- cli/ (drua-cli): thin binary entrypoint, dispatches subcommands
- client/ (drua-client): TUI client library (commands, graphql client)
- server/ (drua-server): server library (routes, auth, GraphQL API)

Single `drua` binary with subcommands:
- `drua server` — start the web server
- `drua tui` — interactive TUI (renamed from dashboard)
- `drua login/logout/status/export/workspace` — client commands
