+++ Makefile
@@ -28,19 +28,19 @@ build-sandbox:
 	cargo build -p sandbox-tool-server
 
 run-server: build-sandbox
-	@PG_CON=$(PG_CON) GITHUB_CLIENT_SECRET=$(GITHUB_CLIENT_SECRET) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) cargo run -p drua-server -- $(ARGS)
+	@PG_CON=$(PG_CON) GITHUB_CLIENT_SECRET=$(GITHUB_CLIENT_SECRET) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) cargo run -p drua-cli -- server $(ARGS)
 
 nix-run-server:
-	@PG_CON=$(PG_CON) GITHUB_CLIENT_SECRET=$(GITHUB_CLIENT_SECRET) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) nix run . -- $(ARGS)
+	@PG_CON=$(PG_CON) GITHUB_CLIENT_SECRET=$(GITHUB_CLIENT_SECRET) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) nix run . -- server $(ARGS)
 
 sdl-rust:
 	SQLX_OFFLINE=true cargo run --bin write_sdl > server/src/graphql/schema.graphql
 
 generate-default-config:
-	SQLX_OFFLINE=true cargo run -q -p drua-server -- dump-default-config > dev/drua.default.yml
+	SQLX_OFFLINE=true cargo run -q -p drua-cli -- server dump-default-config > dev/drua.default.yml
 
 integration-tests: reset-deps
 	DATABASE_URL=$(PG_CON) cargo nextest run
 
 start: reset-deps
-	@PG_CON=$(PG_CON) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) cargo run -p drua-server -- --set oauth.login=dev $(ARGS)
+	@PG_CON=$(PG_CON) ANTHROPIC_API_KEY=$(ANTHROPIC_API_KEY) cargo run -p drua-cli -- server --set oauth.login=dev $(ARGS)
