+++ client/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "drua-client"
+version.workspace = true
+edition.workspace = true
+
+[lib]
+
+[dependencies]
+anyhow = { workspace = true }
+dirs = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+ratatui = { workspace = true }
+crossterm = { workspace = true, features = ["event-stream"] }
+libc = { workspace = true }
+futures = { workspace = true }
