+++ client/src/commands/dashboard.rs
