+++ client/src/commands/login.rs
