+++ client/src/commands/logout.rs
