+++ client/src/commands/status.rs
