+++ client/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod commands;
+pub mod config;
+pub mod graphql;
+pub mod tui;
