+++ client/src/tui/ui/chat_pane.rs
