+++ client/src/tui/ui/sidebar.rs
